ln-simln-jamming: just in time upgrade target htlcs to endorsed
We don't have the ability to set endorsed=1 on the sender with the
current interceptor structure (only intercept forwards, not sender)
so we settle for "just in time" endorsement to mimic the target being
general jammed:
- When a peer forwards the target a htlc:
  - Drop it if the target doesn't have sufficient reputation to have
    it forwarded as endorsed
  - Forward it as endorsed if the target has sufficient reputation

As we're intercepting on the peer, we also need to account for the case
where the peer is the sender and we don't get to intercept on the peer.
This is done by _also_ upgrading all target-forwarded htlcs to endorsed.
The target can force these htlcs being endorsed anyway by jamming their
own channel, so it's not much of a difference.

Of course, as currently specified we'd never upgrade an unendorsed htlc
to endorsed, but this workaround allows us to simulate the target's
peers being general jammed without having to generate real payments to
hold up their resources.