ln-simln-jamming: filter history to allow differential bootstrap
Always bootstrap our network for the full reputation window that we're
interested in, but limit the attacker to the configured bootstrap. These
can of course be the same if configured as such, but allows more
flexible simulations for the case where we're testing an attacker who
hasn't been around as much as the rest of the network.