ln-resource-mgr: fix double lock requirement for add_htlc
Previously we had to lock/unlock our channel map twice to add_htlc,
which could result in inconsistency in our state if we happened to
grab the lock elsewhere in between the two locks.

This commit updates our logic to make use of the newly added inner
forwarding manager to ensure consistency.