multi: allow one slot in congestion bucket if general resources are full
If our general resources are full, there's a good chance that we're
under attack, so we add another bucket that gives honest peers that
do not have sufficient reputation for the protected bucket a chance to
get htlcs through.

Use of these resources is very strict, because we're likely under
attack. We only allow each incoming channel to use on slot in the
congestion bucket, and limit liquidity to be evenly shared between
these slots.