ln-resource-mgr: track average revenue threshold over multiple periods
Previously, we'd just track revenue over a two week decaying average.
This is vulnerable to attacker interference (trying to move thresholds
up or down) or natural fluctuations in traffic.

To smooth this out over time, we track multiple two week periods in a
single decaying average. For convenience, we'll just use the multiplier
that we're using for reputation (ie, track this over 6 months).

This isn't perfect; ideally we'd track values for each 2 week period
and then average them out, but tracking a single average and then
dividing it up into 2 week chunks does an acceptable approximate. In
our checks, there's never more than a 5% difference between this
approach and tracking multiple data points.