ln-simln-jamming: general jam target -> peer channels as well
Previously, we'd general jam all of the target's peers channels in
the direction to the target so that any traffic that reaches the target
is forced to be endorsed. Now, we also general jam the target's channels
in the direction of the peer to trigger use of the target's congestion
resources.