ln-simln-jamming: allow bootstrap for multiple attacker channels
Restricting to a single channel is very specific to a sink attack,
allow multiple so that we can support different attacks.

For now, we're validating that we actually only have one channel as
part of the attack, but once that's pulled out we'll be able to run
with multiple target <-> attacker channels.