ln-simln-jamming: move reputation pair check out of interceptor
Having logic about reputation pairs in the attack interceptor is
very sink-attack specific. Rather surface much more generic information
so that callers can decide how they'd like to check what the current
reputation state is.

Note that a lot of the logic that's bundled into lib will soon be
moved into an attack-specific module, so we don't need to worry about
it getting a bit crowded - it's just there for the same of keeping
commits building.