ln-resource-mgr: return ForwardingOutcome from add_htlc
add_htlc will return the outcome of the proposed forward rather
than the AllocationCheck. add_htlc will internally add the htlc
if it succeeded so it reflects what the forward manager did with
the htlc. Also, here we use the accountable signal on the outgoing
htlc since this signal can be upgraded.