ln-simln-jamming: make resolution timestamp optional for interceptor
We run into "clock drift" issues when we pick our timestamp before
acquiring a hot path lock on network_nodes. We only need to set this
value for bootstraps when we're replaying them, so we make this field
optional and set it once the lock has been acquired.

We only practically hit this issue when dealing with large clock
speedups, because the amount of time between lock acquisitions is
multiplied by the speedup, which can cause issues with the decyaing
averages we use which are only tolerant of nanosecond differences.