ln-simln-jamming: redefine run_custom_actions to run_attack.
Core of the attack is now in run_attack and removed the
simulation_completed polling check. It is now up to each attack
impl to define the shutdown conditions and return from run_attack.
In the case of Sink impl, it mostly consists of moving the polling
check to run_attack.