ln-simln-jamming/refactor: remove dyn clock type from implementation
We've got clock generics/dyns all over the place, but we only ever
use a SimulationClock. Simplify this code to just accept the concrete
type.

If we need to mock out time, we can revert this change to allow mocking
the clock, but we've achieved reasonable test coverage without it so
far so seem unlikely to need it.